# Add EDI's `file_declaration` struct definition and implement `ediReader.getUnprocessedRawSeg`