--- a/extensions/omniv21/fileformat/edi/bench_test.go
+++ b/extensions/omniv21/fileformat/edi/bench_test.go
@@ -1,0 +1,389 @@
+package edi
+
+import (
+	"encoding/json"
+	"io"
+	"strings"
+	"testing"
+)
+
+// Adding a benchmark for rawSeg operation to ensure there is no alloc:
+// BenchmarkRawSeg-8   	81410766	        13.9 ns/op	       0 B/op	       0 allocs/op
+func BenchmarkRawSeg(b *testing.B) {
+	rawSegName := "test"
+	rawSegData := []byte("test data")
+	r := ediReader{
+		unprocessedRawSeg: newRawSeg(),
+	}
+	for i := 0; i < b.N; i++ {
+		r.resetRawSeg()
+		r.unprocessedRawSeg.valid = true
+		r.unprocessedRawSeg.name = rawSegName
+		r.unprocessedRawSeg.raw = rawSegData
+		r.unprocessedRawSeg.elems = append(
+			r.unprocessedRawSeg.elems, rawSegElem{1, 1, rawSegData[0:4]}, rawSegElem{2, 1, rawSegData[5:]})
+	}
+}
+
+// Adding a benchmark for stack operation to ensure there is no alloc:
+// BenchmarkGrownShrinkStack-8    	12901227	        89.0 ns/op	       0 B/op	       0 allocs/op
+func BenchmarkGrownShrinkStack(b *testing.B) {
+	r := ediReader{
+		stack: newStack(),
+	}
+	for i := 0; i < b.N; i++ {
+		for j := 0; j < 20; j++ {
+			r.growStack(stackEntry{})
+		}
+		for r.shrinkStack() != nil {
+		}
+	}
+}
+
+const (
+	benchInputNoCompNoReleaseChar = `
+ISA*00*          *00*          *02*CPC            *ZZ*00602679321    *191103*1800*U*00401*000001644*0*P*>
+GS*QM*CPC*00602679321*20191103*1800*000001644*X*004010
+ST*214*000000001
+B10*4343638097845589              *4343638097845589              *CPCC
+L11*4343638097845589*97
+L11*0000*86
+N1*SF*00602679321
+N1*ST*The Rock
+N3*264040 DONUT-TOWN 473*
+N4*HAMMER*AB*T0C1Z0*CA
+LX*000001
+AT7*XB*NS***20191103*1534*
+MS1*WRENCH*ON*CA
+AT8*G*K*4
+SE*0000000013*000000001
+ST*214*000000002
+B10*4343638098050296              *4343638098050296              *CPCC
+L11*4343638098050296*97
+L11*0156*86
+N1*SF*00602679321
+N1*ST*The Terminator
+N3*PO BOX 12345 RPO 21ST AVE*MARKET
+N4*DRILL*BC*V5L0B3*CA
+LX*000001
+AT7*AH*AG***20191102*1625*PT
+MS1*DRILL*BC*CA
+AT8*G*K*0
+SE*0000000013*000000002
+ST*214*000000003
+B10*4343638931638575              *4343638931638575              *CPCC
+L11*3001100072*97
+L11*1303*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 42*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*CA*BT***20191102*1752*PT
+MS1*SAW*BC*CA
+AT8*G*K*0
+SE*0000000013*000000003
+ST*214*000000004
+B10*4343638098146166              *4343638098146166              *CPCC
+L11*4343638098146166*97
+L11*1498*86
+N1*SF*00602679321
+N1*ST*The Joker
+N3*RR 1*
+N4*SWITCH*AB*T0C0J0*CA
+LX*000001
+AT7*D1*NS***20191102*1648*MT
+MS1*SWITCH*AB*CA
+L11**CI
+AT8*G*K*15
+SE*0000000014*000000004
+ST*214*000000005
+B10*4343638098181877              *4343638098181877              *CPCC
+L11*4343638098181877*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*The Autobot
+N3*PO BOX 314*159 26 ST-GRINDER
+N4*SCREWDRIVER*SK*S0M0E0*CA
+LX*000001
+AT7*AF*NS***20191103*1509*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*1
+SE*0000000013*000000005
+ST*214*000000006
+B10*4343638098181891              *4343638098181891              *CPCC
+L11*4343638098181891*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*The Hobbit
+N3*PO BOX 42*
+N4*MITERSAW*SK*S0M1C0*CA
+LX*000001
+AT7*AF*NS***20191103*1509*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*19
+SE*0000000013*000000006
+ST*214*000000007
+B10*4343638098181921              *4343638098181921              *CPCC
+L11*4343638098181921*97
+L11*0100*86
+N1*SF*00602679321
+N1*ST*The Avengers
+N3*PO BOX 135*HOUSE 9 IMPACTDRILL
+N4*WETVAC*MB*R0A1E0*CA
+LX*000001
+AT7*XB*NS***20191103*1131*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*0
+SE*0000000013*000000007
+ST*214*000000008
+B10*4343638098186995              *4343638098186995              *CPCC
+L11*4343638098186995*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*The Hulk
+N3*PO BOX 000*
+N4*PRESSUREWASHER*SK*S0K2L0*CA
+LX*000001
+AT7*AF*NS***20191103*1509*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*0
+SE*0000000013*000000008
+ST*214*000000009
+B10*4343638151403540              *4343638151403540              *CPCC
+L11*5653512*97
+L11*0175*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*AF*NS***20191103*1125*MT
+MS1*LASERLEVEL*AB*CA
+AT8*G*K*0
+SE*0000000013*000000009
+ST*214*000000010
+B10*4343638151403540              *4343638151403540              *CPCC
+L11*5653512*97
+L11*0100*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*XB*NS***20191103*0853*MT
+MS1*LASERLEVEL*AB*CA
+AT8*G*K*0
+SE*0000000013*000000010
+ST*214*000000011
+B10*4343638316026577              *4343638316026577              *CPCC
+L11*300521101*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*AF*NS***20191103*0837*MT
+MS1*LASERLEVEL*AB*CA
+AT8*G*K*0
+SE*0000000013*000000011
+ST*214*000000012
+B10*4343638316026577              *4343638316026577              *CPCC
+L11*300521101*97
+L11*0405*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*AP*AG***20191103*0815*MT
+MS1*LASERLEVEL*AB*CA
+AT8*G*K*0
+SE*0000000013*000000012
+ST*214*000000013
+B10*4343638316026577              *4343638316026577              *CPCC
+L11*300521101*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*AF*NS***20191102*1555*PT
+MS1*NAILGUN*BC*CA
+AT8*G*K*0
+SE*0000000013*000000013
+ST*214*000000014
+B10*4343638672340607              *4343638672340607              *CPCC
+L11*300633104*97
+L11*0405*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*AP*AG***20191103*0733*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*0
+SE*0000000013*000000014
+ST*214*000000015
+B10*4343638098171441              *4343638098171441              *CPCC
+L11*4343638098171441*97
+L11*0100*86
+N1*SF*00602679321
+N1*ST*The Dragon Ball
+N3*PO BOX 314*159 2 ST
+N4*WOODGLUE*AB*T0K2E0*CA
+LX*000001
+AT7*XB*NS***20191103*0904*MT
+MS1*LASERLEVEL*AB*CA
+AT8*G*K*16
+SE*0000000013*000000015
+ST*214*000000016
+B10*4343638098171472              *4343638098171472              *CPCC
+L11*4343638098171472*97
+L11*0410*86
+N1*SF*00602679321
+N1*ST*The X-men
+N3*PO BOX 314*1592 WIREBOX PL
+N4*DRYWALL*SK*S0L0P0*CA
+LX*000001
+AT7*AF*NS***20191103*1509*CS
+MS1*PAINTBRUSH*MB*CA
+AT8*G*K*3
+SE*0000000013*000000016
+ST*214*000000017
+B10*4343638098176088              *4343638098176088              *CPCC
+L11*4343638098176088*97
+L11*0405*86
+N1*SF*00602679321
+N1*ST*The Matrix
+N3*PO BOX 314*159 UP ST
+N4*FUSE*SK*S0C1S0*CA
+LX*000001
+AT7*AP*AG***20191103*0631*CS
+MS1*SHEETROCK*SK*CA
+AT8*G*K*0
+SE*0000000013*000000017
+ST*214*000000018
+B10*4343638458862606              *4343638458862606              *CPCC
+L11*3001570460*97
+L11*0104*86
+N1*SF*00602679321
+N1*ST*
+N3*PO BOX 1122*
+N4*WRENCH*ON*N5Y5W3*CA
+LX*000001
+AT7*XB*NS***20191103*1728*ET
+MS1*PVCPIPE*ON*CA
+AT8*G*K*0
+SE*0000000013*000000018
+ST*214*000000019
+B10*4343638098196574              *4343638098196574              *CPCC
+L11*4343638098196574*97
+L11*0104*86
+N1*SF*00602679321
+N1*ST*The Beast
+N3*PO BOX 314*1592 HWY 551
+N4*WIRECUTTER*ON*P0P1S0*CA
+LX*000001
+AT7*XB*NS***20191103*1618*ET
+MS1*TORCH*ON*CA
+AT8*G*K*1
+SE*0000000013*000000019
+GE*000019*000001644
+IEA*00001*000001644
+`
+
+	benchDeclNoCompNoReleaseCharJSON = `
+{
+        "segment_delimiter": "\n",
+        "element_delimiter": "*",
+        "segment_declarations": [
+            {
+                "name": "ISA",
+                "child_segments": [
+                    {
+                        "name": "GS",
+                        "child_segments": [
+                            {
+                                "name": "scanInfo", "type": "segment_group", "min": 0, "max": -1, "is_target": true,
+                                "child_segments": [
+                                    { "name": "ST" },
+                                    { "name": "B10", "elements": [ { "name": "shipmentIdentificationNumber", "index": 2 } ] },
+                                    { "name": "L11" },
+                                    { "name": "L11" },
+                                    { "name": "N1" },
+                                    { "name": "N1" },
+                                    { "name": "N3" },
+                                    {
+                                        "name": "N4",
+                                        "elements": [
+                                            { "name": "cityName", "index": 1 },
+                                            { "name": "provinceCode", "index": 2 },
+                                            { "name": "postalCode", "index": 3 },
+                                            { "name": "countryCode", "index": 4 }
+                                        ]
+                                    },
+                                    { "name": "LX" },
+                                    {
+                                        "name": "AT7",
+                                        "elements": [
+                                            { "name": "shipmentStatusCode", "index": 1 },
+                                            { "name": "shipmentStatusReasonCode", "index": 2 },
+                                            { "name": "date", "index": 5 },
+                                            { "name": "time", "index": 6 },
+                                            { "name": "timeCode", "index": 7 }
+                                        ]
+                                    },
+                                    {
+                                        "name": "MS1",
+                                        "elements": [
+                                            { "name": "cityName", "index": 1 },
+                                            { "name": "provinceCode", "index": 2 },
+                                            { "name": "countryCode", "index": 3 }
+                                        ]
+                                    },
+                                    { "name": "L11", "min": 0 },
+                                    { "name": "AT8" },
+                                    { "name": "SE" }
+                                ]
+                            }
+                        ]
+                    },
+                    { "name": "GE" }
+                ]
+            },
+            { "name": "IEA" }
+        ]
+    }`
+)
+
+var (
+	benchDeclNoCompNoReleaseChar = func() *fileDecl {
+		var fd fileDecl
+		err := json.Unmarshal([]byte(benchDeclNoCompNoReleaseCharJSON), &fd)
+		if err != nil {
+			panic(err)
+		}
+		return &fd
+	}()
+)
+
+// BenchmarkGetUnprocessedRawSeg_NoCompNoReleaseChar-8   	    8971	    114907 ns/op	   48024 B/op	     747 allocs/op
+func BenchmarkGetUnprocessedRawSeg_NoCompNoReleaseChar(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		reader := NewReader("test", strings.NewReader(benchInputNoCompNoReleaseChar), benchDeclNoCompNoReleaseChar)
+		for {
+			_, err := reader.getUnprocessedRawSeg()
+			if err == io.EOF {
+				break
+			}
+			if err != nil {
+				b.FailNow()
+			}
+			reader.resetRawSeg()
+		}
+	}
+}