The implementation tries to minimize allocation/reallocation induced by conversions among `string<->[]rune<->[]byte`.
Benchmark shows significant improvement (~4x to ~5x in latency reduction in general and ~20x in memory alloc reduction in most popular scenario that input has no comp-delim nor release-char) 
- most popular scenario: no comp-delim and no release-char:
```
old: BenchmarkGetUnprocessedRawSeg_NoCompNoReleaseChar-8   3000  389955 ns/op  558626 B/op  3024 allocs/op
new: BenchmarkGetUnprocessedRawSeg_NoCompNoReleaseChar-8  10000  103945 ns/op   27080 B/op   515 allocs/op
```

- most complex scenario: comp-delim + release-char:
```
old: BenchmarkGetUnprocessedRawSeg_WithCompAndRelease-8    3000  402042 ns/op  397383 B/op  1857 allocs/op
new: BenchmarkGetUnprocessedRawSeg_WithCompAndRelease-8   14174   84115 ns/op  102880 B/op   385 allocs/op
```